Add metatable support to generated messages.
This includes letting .new take initialization data in order to preserve workflows in which a
message is created from a pure data table.

Also,  WKTs' jsonDecode methods now take and use a constructor, so that they can return the
metatable object rather than just a data table.

This makes it possible to write e.g. my_duration:encode rather than having to write duration.Duration.encode(my_duration).  This is always convenient, but it's necessary for generic code like Any.pack.

Finally, generated code for WKTs is included in src/ (and can be regenerated with `just generate-wkts`).  This is needed in order for the implementation of the WKTs' special JSON decoding (e.g. a Struct containing a Value needs to be able to call Value.decodeJson).